fix: removed error module
Error now is present in the required module itself
