+++ src/main.rs
@@ -1,7 +1,6 @@
 mod cli;
 mod config;
 mod deps;
-mod error;
 mod util;
 
 fn main() -> std::process::ExitCode {
